postgres_to_es/requirements.txt
python-dotenv==1.0
psycopg2-binary==2.9.6
elasticsearch==8.7.0
pydantic==1.10.7